working comments and posts

+++ ClientApp/src/app/feed/feed.component.ts
@@ -12,6 +12,8 @@ export class FeedComponent implements OnInit {
   url: string;
   writtenPosts: Post[];
   profileUrl: string;
+  comments: Map<number, Comment[]>;
+  baseUrl: string;
 
   constructor(
     private burritoService: BurritoService,
@@ -24,6 +26,8 @@ export class FeedComponent implements OnInit {
     this.writer = this.burritoService.getUsername();
     this.url = baseUrl + 'api/post';
     this.profileUrl = baseUrl + 'profile';
+    this.baseUrl = baseUrl;
+    this.comments = new Map<number, Comment[]>();
 
     console.log(this.writer);
   
@@ -36,13 +40,18 @@ export class FeedComponent implements OnInit {
   share(content) {
     console.log(content);
 
-    this.http.post(this.url, {
-      "id": 2,
-      "username": localStorage.getItem("logged"),
-      "content": content
-    }).subscribe((result) => {
-      console.log("posted");
-    });
+    if (localStorage.getItem("logged")) {
+      this.http.post(this.url, {
+        "id": 2,
+        "username": localStorage.getItem("logged"),
+        "content": content
+      }).subscribe(result => {
+        console.log("posted");
+      });
+    }
+    else {
+      console.error("Not logged");
+    }
 
     window.location.reload();
   }
@@ -50,13 +59,44 @@ export class FeedComponent implements OnInit {
   retrievePosts() {
     this.http.get<Post[]>(this.url).subscribe(result => {
       this.writtenPosts = result;
+    
+      this.writtenPosts.forEach(it => {
+        this.http.get<Comment[]>(this.baseUrl + `api/comment/post/${it.id}`).subscribe(result => {
+          console.log(result);
+          this.comments.set(it.id, result);
+        });
+      });
+
+      console.log(this.comments);
     });
   }
 
   delete(postId, postUsername) {
-    this.http.delete(this.url + `/${postId}?username=${postUsername}`).subscribe(result => {
-      console.log("Deleted the post");
-    });
+    if (localStorage.getItem("logged")) {
+      this.http.delete(this.url + `/${postId}?username=${postUsername}`).subscribe(result => {
+        console.log("Deleted the post");
+      });
+    }
+    else {
+      console.error("Not logged");
+    }
+
+    window.location.reload();
+  }
+
+  writeComment(postid, content) {
+    if (localStorage.getItem("logged")) {
+      this.http.post(this.baseUrl + 'api/comment', {
+        "postID": postid,
+        "username": localStorage.getItem("logged"),
+        "content": content
+      }).subscribe(result => {
+        console.log("commented");
+      });
+    }
+    else {
+      console.error("Not logged");
+    }
 
     window.location.reload();
   }
@@ -64,6 +104,14 @@ export class FeedComponent implements OnInit {
 
 interface Post {
   id: number,
+  username: string,
+  content: string,
+  date: string
+}
+
+interface Comment {
+  id: number,
+  postID: number,
   username: string,
   content: string,
   date: string
